+++ app/Http/Controllers/SupplierController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Supplier;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Validation\Rule;
 
 class SupplierController extends Controller
@@ -15,6 +16,10 @@ class SupplierController extends Controller
      */
     public function index()
     {
+        if (Auth::user()->role !== "admin") {
+            return View("403");
+        }
+
         $suppliers = Supplier::latest()->get();
 
         return view("suppliers.index", compact("suppliers"));
@@ -27,6 +32,10 @@ class SupplierController extends Controller
      */
     public function create()
     {
+        if (Auth::user()->role !== "admin") {
+            return View("403");
+        }
+
         return view("suppliers.create");
     }
 
@@ -38,6 +47,10 @@ class SupplierController extends Controller
      */
     public function store(Request $request)
     {
+        if (Auth::user()->role !== "admin") {
+            return View("403");
+        }
+
         $validated = $request->validate([
             "name" => ["required", "unique:suppliers"],
             "phone" => ["required", "string"],
@@ -68,6 +81,10 @@ class SupplierController extends Controller
      */
     public function edit(Supplier $supplier)
     {
+        if (Auth::user()->role !== "admin") {
+            return View("403");
+        }
+
         return view("suppliers.edit", compact("supplier"));
     }
 
@@ -80,6 +97,10 @@ class SupplierController extends Controller
      */
     public function update(Request $request, Supplier $supplier)
     {
+        if (Auth::user()->role !== "admin") {
+            return View("403");
+        }
+
         $validated = $request->validate([
             "name" => [ "required", "string", Rule::unique('suppliers')->ignore($supplier->name, 'name') ],
             "phone" => ["required", "string"],
@@ -102,6 +123,10 @@ class SupplierController extends Controller
      */
     public function destroy(Supplier $supplier)
     {
+        if (Auth::user()->role !== "admin") {
+            return View("403");
+        }
+
         $supplier->delete();
 
         return redirect()->route("suppliers.index");
